+++ apps/browser-extension/background/index.js
@@ -1,0 +1,10 @@
+import "./init.js";
+import "../assets/browser-polyfill.min.js";
+import "../editor/youtube-auth.js";
+import "../editor/youtube-api.js";
+import "../editor/storage-service.js";
+import "../editor/utils.js";
+import "../editor/video-service.js";
+import "../editor/supabase-sync.js";
+import "./background.js";
+